Flesh out reference runtime implementation, reach parity with minimal runtime (#155)
* Cleanup bootstrap and hookup dynlink context discovery in monitor.

* Adding unwinding support for reference runtime.

* Fix libunwind for reference runtime.

* Fix minalign magic num.

* Fix early slot reservation.

* Update, cleanup, and first pass of object runtime.

* Implement first-pass of threading runtime.

* Fix comments, run clippy.

+++ src/lib/twizzler-abi/src/syscall/handle.rs
@@ -109,7 +109,7 @@ pub fn sys_new_handle(
             &[hi, lo, handle_type as u64, flags.bits()],
         )
     };
-    convert_codes_to_result(code, val, |c, _| c != 0, |_, v| v as u64, justval)
+    convert_codes_to_result(code, val, |c, _| c != 0, |_, v| v, justval)
 }
 
 /// Unbind an object from handle status.
